pyproject.toml
[tool.poetry]
name = "portal-ai"
version = "0.0.1"
description = "Portal-AI is a tool for deploying AWS infrastructure to run GPU accelerated workflows and websites for LLMs."
authors = ["Schlafdata <john@schlafdata.com>"]
readme = "README.md"
license = ""

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pyyaml = "6.0.1"
jinja2 = "3.1.3"
pytest = "8.1.1"
pytest-mock = "3.14.0"
boto3= "1.34.74"
aws2-wrap = "1.4.0"
kubernetes = "29.0.0"
psycopg2-binary = "2.9.9"
python-dotenv = "1.0.1"
hvac = "2.1.0"

[tool.poetry.dev-dependencies]
# Add your development dependencies here. For example:
# pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Homepage" = "https://github.com/jschlafdata/portal-ai"